_archive/cockpit-legacy/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project]
name = "lanecore-backend"
version = "0.1.0"
description = "LaneCore AI Backend — Bauplan-Analyse & Angebotsautomatisierung"
requires-python = ">=3.12"

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # Claude API
    "anthropic>=0.40.0",

    # PDF Processing
    "pypdf>=4.0.0",
    "pdfplumber>=0.11.0",
    "Pillow>=10.0.0",       # PDF → Image für Vision-API
    "pdf2image>=1.17.0",    # Hochwertige PDF-zu-Bild-Konvertierung

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",

    # Validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",

    # Storage (S3)
    "boto3>=1.35.0",
    "aioboto3>=13.0.0",

    # Auth (Clerk JWT validation)
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",

    # Excel Export
    "openpyxl>=3.1.0",

    # PDF Export (Angebote)
    "reportlab>=4.1.0",

    # Utilities
    "python-multipart>=0.0.12",  # File uploads
    "structlog>=24.0.0",          # Structured logging
    "tenacity>=8.5.0",            # Retry logic für Claude API

    # Observability & Subscriptions
    "sentry-sdk[fastapi,sqlalchemy]>=2.0.0",  # Error tracking
    "stripe>=11.0.0",                          # Subscription billing
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",       # TestClient
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "factory-boy>=3.3.0",  # Test Fixtures
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
